Renames Lepy to PyLego
