+++ setup.py
@@ -1,4 +1,4 @@
-"""Build lepy."""
+"""Build pylego."""
 
 import os
 import subprocess
@@ -9,7 +9,7 @@ from setuptools.command.build_py import build_py as build_py_orig
 
 def build_go_library():
     """Build the lego application into a shared .so file."""
-    os.chdir("src/lepy")
+    os.chdir("src/pylego")
     subprocess.check_call(["go", "build", "-o", "lego.so", "-buildmode=c-shared", "lego.go"])
     os.chdir("../..")
