--- src/lepy/__init__.py
@@ -1,1 +1,0 @@
-from .lepy import run_lego_command  # noqa: F401, D104
