+++ src/pylego/__init__.py
@@ -1,0 +1,1 @@
+from .pylego import run_lego_command  # noqa: F401, D104
